+++ c_src/reformat.c
@@ -893,8 +893,9 @@ int main(int argc, char** argv)
     test_reformat(src, dst, NUM_FRAMES, NUM_REPEATS);
     t1 = time_tick();
     printf("time = %lus+%luus, %f reformat/s\n",
-	   (t1-t0) / 1000000, (t1-t0) % 1000000,
-	   ((float)NUM_REPEATS / ((t1-t0)/1000000.0)));
+	   ((unsigned long)(t1-t0)) / 1000000,
+	   ((unsigned long)(t1-t0)) % 1000000,
+	   ((float)NUM_REPEATS / (((unsigned long)(t1-t0))/1000000.0)));
     exit(0);
 }
